Lifted shared cache code into trait.
