+++ src/Drivers/ExchangeRatesApILegacy/ExchangeRatesApiLegacyDriver.php
@@ -4,6 +4,7 @@ namespace AshAllenDesign\LaravelExchangeRates\Drivers\ExchangeRatesApILegacy;
 
 use AshAllenDesign\LaravelExchangeRates\Classes\CacheRepository;
 use AshAllenDesign\LaravelExchangeRates\Classes\Validation;
+use AshAllenDesign\LaravelExchangeRates\Concerns\InteractsWithCache;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\ExchangeRateException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidCurrencyException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidDateException;
@@ -15,6 +16,8 @@ use GuzzleHttp\Exception\GuzzleException;
 
 class ExchangeRatesApiLegacyDriver implements ExchangeRateDriver
 {
+    use InteractsWithCache;
+
     /**
      * The object used for making requests to the currency
      * conversion API.
@@ -319,54 +322,5 @@ class ExchangeRatesApiLegacyDriver implements ExchangeRateDriver
         }
 
         return $conversions;
-    }
-
-    /**
-     * Determine whether if the exchange rate should be
-     * cached after it is fetched from the API.
-     *
-     * @param  bool  $shouldCache
-     * @return $this
-     */
-    public function shouldCache(bool $shouldCache = true): self
-    {
-        $this->shouldCache = $shouldCache;
-
-        return $this;
-    }
-
-    /**
-     * Determine whether if the cached result (if it
-     * exists) should be deleted. This will force
-     * a new exchange rate to be fetched from
-     * the API.
-     *
-     * @param  bool  $bustCache
-     * @return $this
-     */
-    public function shouldBustCache(bool $bustCache = true): self
-    {
-        $this->shouldBustCache = $bustCache;
-
-        return $this;
-    }
-
-    /**
-     * Attempt to fetch an item (more than likely an
-     * exchange rate) from the cache. If it exists,
-     * return it. If it has been specified, bust
-     * the cache.
-     *
-     * @param  string  $cacheKey
-     * @return mixed
-     */
-    private function attemptToResolveFromCache(string $cacheKey)
-    {
-        if ($this->shouldBustCache) {
-            $this->cacheRepository->forget($cacheKey);
-            $this->shouldBustCache = false;
-        } elseif ($cachedValue = $this->cacheRepository->getFromCache($cacheKey)) {
-            return $cachedValue;
-        }
     }
 }
